% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/semdrw.R
\name{semdrw}
\alias{semdrw}
\title{Launch semdrw Interface}
\usage{
semdrw()
}
\value{
Nothing
}
\description{
semdrw() loads interactive user interface built using R shiny.
}
\details{
Start semdrw

The interactive user interface is to provide an easy way for people who are learning sem  and writing research paper. Includes example data for testing out a few example analysis.
}
\examples{
\dontrun{
library(shiny)
semdrw()
}
}
\keyword{semdrw}
